requirements.txt
fastapi
uvicorn
jinja2
python-multipart
transformers
torchaudio
soundfile
python-dotenv